+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 Changes:
 
 - fix: missing enchants listing enchantments that will conflict with an enchantment already present
+
+New Additions:
+
+- feat: added Not maxed enchantments line to the tooltip